Attempt to fix Dependabot prefixes
By default, Dependabot apparently tries to match the style of commits in
the repository when creating commits/PR titles. We have a number of
early commits that do `<component>: <message>`, so it seemingly tried to
match that in #27, which goes against the pr-title "linter" we use on
Connect projects.

Not sure this will 100% fix the issue, but no real other way to find out
other than trying this out and recreating the PR.

If we feel strongly about another commit style, we could drop the
linter, but I'd prefer to stay consistent if we can.

[1]: https://docs.github.com/en/code-security/dependabot/working-with-dependabot/dependabot-options-reference#commit-message--

Signed-off-by: Stefan VanBuren <svanburen@buf.build>