Rename plugin to protoc-gen-connect-python
Makes the plugin name more consistent with the other Connect RPC plugins
(and probably most protobuf plugins in general).

No strong feeling about this, as I see that e.g. the [mypy-protobuf
plugins][1] use underscores as well, so if this is somehow easier to use
in Python I think it's fine to not rename them.

Resolves #19.

[1]: https://github.com/nipunn1313/mypy-protobuf

Signed-off-by: Stefan VanBuren <svanburen@buf.build>